Separate wasm integration-test to multiple test files.
